requirements.txt
# This file was generated by inspecting the 'py37' conda environment.
# It reflects the exact versions used in the original experiments.

# Core Frameworks
# Python Version: 3.7.12
# PyTorch was installed from the 'pytorch' channel with CUDA 11.1 support
torch==1.9.0
# DGL was installed from the 'dglteam' channel with CUDA 11.1 support
dgl-cuda11.1==0.9.1
torchvision==0.10.0

# Key Libraries
numpy==1.21.6
prettytable==2.1.0
scikit-learn==1.0.2
pandas==1.3.5
matplotlib==3.5.3
